refactor: move db predicate conversion to its own module (#664)
